+++ mitra_api/src/mastodon_api/statuses/mod.rs
@@ -1,3 +1,4 @@
 pub mod helpers;
 pub mod types;
+pub mod utils;
 pub mod views;