+++ CHANGELOG.md
@@ -13,6 +13,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - Added `/api/v1/favourites` API endpoint.
 - Added `/api/v1/preferences` API endpoint.
 - Added shared default post visibility setting.
+- Added shared default post language setting.
 
 ### Changed